+++ src/kiara/_frkl/__init__.py
@@ -1,4 +1,0 @@
-# -*- coding: utf-8 -*-
-from typing import Any, Dict
-
-build_properties: Dict[str, Any] = {"resources": [], "hidden_imports": []}